separate daemon from controller, make daemon the main() entrypoint
